chore: move components to Feature
