+++ src/components/index.ts
@@ -1,12 +1,13 @@
 import OACodeBlock from './Common/OACodeBlock.vue'
 import OAFooter from './Common/OAFooter.vue'
 import OAHeading from './Common/OAHeading.vue'
-import OAInfo from './Common/OAInfo.vue'
-import OAIntroduction from './Common/OAIntroduction.vue'
 import OALocaleSelect from './Common/OALocaleSelect.vue'
 import OAMarkdown from './Common/OAMarkdown.vue'
-import OAServers from './Common/OAServers.vue'
-import OAOperation from './Operation/OAOperation.vue'
+import OAInfo from './Feature/OAInfo.vue'
+import OAIntroduction from './Feature/OAIntroduction.vue'
+import OAOperation from './Feature/OAOperation.vue'
+import OAServers from './Feature/OAServers.vue'
+import OASpec from './Feature/OASpec.vue'
 import OAParameter from './Parameter/OAParameter.vue'
 import OAParameters from './Parameter/OAParameters.vue'
 import OAPath from './Path/OAPath.vue'
@@ -22,7 +23,6 @@ import OACodeSamples from './Sample/OACodeSamples.vue'
 import OASchemaPropertyAttributes from './Schema/OASchemaPropertyAttributes.vue'
 import OASchemaTabs from './Schema/OASchemaTabs.vue'
 import OASecurity from './Security/OASecurity.vue'
-import OASpec from './Spec/OASpec.vue'
 import OASpecSkeleton from './Spec/OASpecSkeleton.vue'
 import OATryItButton from './Try/OATryItButton.vue'
 import OATryWithVariables from './Try/OATryWithVariables.vue'
